requirements.txt
streamlit
pydantic
PyMuPDF
pdfplumber
faiss-cpu
sentence-transformers
groq
networkx
pandas
numpy
requests
